requirements.txt
--extra-index-url https://pypi.fury.io/Jmtu27PFLB2CHCyHa4LF/aerovane/
amqp==1.4.7
anyjson==0.3.3
apache-libcloud===0.20.4-aerovane
Babel==2.2.0
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.21
boto==2.38.0
celery==3.1.23
defusedxml==0.4.1
dj-database-url==0.3.0
Django==1.8.13
django-admin-bootstrapped==2.5.6
django-allauth==0.24.0
django-annoying==0.8.4
django-appconf==1.0.2
django-bootstrap3==6.2.2
django-compressor==2.0
django-extensions==1.5.9
django-log-request-id==1.1.0
django-mailgun==0.8.0
django-polymorphic==0.7.2
-e git+https://github.com/karanlyons/django-save-the-change.git@e48502d2568d76bd9c7093f4c002a5b0061bc468#egg=django_save_the_change
django-simple-history==1.7.0
django-storages==1.1.8
django-storages-redux==1.3
django-transaction-hooks==0.2
flower==0.9.1
gunicorn==19.3.0
haikunator==1.0.1
httplib2==0.9.2
kombu==3.0.35
librato-metrics==0.8.6
librato-python-web==0.40.6
newrelic==2.64.0.48
oauth2client==1.5.1
oauthlib==1.0.3
psycopg2==2.6.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycrypto==2.6.1
python3-openid==3.0.9
pytz==2015.7
rcssmin==1.0.6
requests==2.8.1
requests-oauthlib==0.5.0
rjsmin==1.0.12
rsa==3.2.3
simplejson==3.8.1
six==1.10.0
tornado==4.3
uritemplate==0.6
whitenoise==3.0
